requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
apscheduler
argon2-cffi
pyjwt
python-ulid
slowapi
